web-based/requirements.txt
asgiref==3.9.1
bottle==0.13.4
cffi==1.17.1
clipboard==0.0.4
clr_loader==0.2.7.post0
Django==5.2.4
django-cors-headers==4.7.0
keyboard==0.13.5
pillow==11.3.0
proxy_tools==0.1.0
pycparser==2.22
pyperclip==1.9.0
pythonnet==3.0.5
pywebview==5.4
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
